shorten attribution notices -> legal notices
It's pretty, uh, dry reading, and takes up space. I *would* just move it
to settings, but then the link wouldn't be accessible to logged-out
users. I *could* make the footer different for logged-out users, but
that would be confusing. And I've been working on a marketing site
seperate from this repo that's just a static site. Once I finish that,
then I should be able to just move this to settings. It might also be
able to instead make the link a source code comment, but it'll be easier
to just move it to settings later.

The old wording does make a bit more sense: the notices are for
attribution, and they are required by the licence, not law. Meh. Too
long.